+++ src/tools/the-beast-code-gen/src/client.ts
@@ -7,16 +7,13 @@ export class MEEditorClient {
   constructor () {
     this.ws=new WebSocket("ws://localhost:1243");
     this.ws.onopen=() => {
-      // if(typeOfRun=='created from editor') {
-      //   console.log(`%c <signal> ${name}`, LOG_FUNNY_ARCADE);
-      //   // var o: any={
-      //   //   action: "watch",
-      //   //   name: name,
-      //   //   path: ''
-      //   // };
-      //   // o=JSON.stringify(o);
-      //   // this.ws.send(o);
-      // }
+      console.log(`%c <signal> list`, LOG_FUNNY_ARCADE);
+      var o: any={
+        action: "list",
+        path: ''
+      };
+      o=JSON.stringify(o);
+      this.ws.send(o);
       console.log("%c[CODE CREATOR][WS OPEN]", LOG_FUNNY_ARCADE);
       document.dispatchEvent(new CustomEvent("code-creator-ready", {}));
     };
@@ -64,14 +61,16 @@ export class MEEditorClient {
       this.ws.send(o);
     });
 
-    // document.addEventListener('delete-shader-graph', (e: any) => {
-    //   console.info('%cDelete shader-graph <signal>', LOG_FUNNY_ARCADE);
-    //   let o: any = {
-    //     action: "delete-shader-graph",
-    //     name: e.detail
-    //   };
-    //   o = JSON.stringify(o);
-    //   this.ws.send(o);
-    // });
+    document.addEventListener('nav-folder', (e: any) => {
+      console.info('nav-folder <signal>');
+      let o={
+        action: "nav-folder",
+        name: e.detail.name,
+        rootFolder: e.detail.rootFolder
+      };
+      o=JSON.stringify(o);
+      this.ws.send(o);
+    });
+
   }
 }